fix: make get_supplier_part only return matches if the company match
I had a bug starting due two suppliers (TME & TI) both use the same SKU,
then it creates the TI supplier part everything is fine,
after that it creates the TME supplier part, does a SKU search only find
one single result and return it as the existing supplier part even tho
this is TI's supplier part not TME's.
Then TME would proceed to update the TI supplier part with the TME info.

Making a last one wins and making it impossible for boths to be listed for the same part.